% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_colors.R
\name{emodnet_colors}
\alias{emodnet_colors}
\title{Load colors of EMODnet}
\usage{
emodnet_colors()
}
\value{
list
}
\description{
This function returns a list with the six colors palette of EMODnet Biology, and white
}
\examples{
emodnet_colors()$blue
emodnet_colors()$darkblue
}
