% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_map_logo.R
\name{emodnet_map_logo}
\alias{emodnet_map_logo}
\title{Add the EMODnet logo to your ggplot}
\usage{
emodnet_map_logo(
  plot,
  path = NULL,
  gravity = "northeast",
  offset = "+0+0",
  color = TRUE,
  ...
)
}
\arguments{
\item{plot}{ggplot}

\item{path}{character}

\item{gravity}{character}

\item{offset}{character}

\item{color}{boolean}

\item{...}{params to be passed to ggplot2::ggsave()}
}
\value{
png
}
\description{
Add the EMODnet logo to your ggplot
}
\examples{
aphiaid <- 107451
specname <- "Eriocheir sinensis"
Esgrid <- sf::st_read(paste0("http://geo.vliz.be/geoserver/wfs/ows?", "service=WFS&version=1.3.0&",
"request=GetFeature&", "typeName=Dataportal\%3Aeurobis_grid_1d",
"-obisenv&", "viewParams=aphiaid\%3A", aphiaid, "&", "outputFormat=json", "&maxFeatures=10"))
plot <- emodnet_map_plot(Esgrid, fill = Esgrid$RecordCount,  title = specname,
subtitle = paste("Aphiaid =",aphiaid), legend = "Abundance",plot_polygon_border=TRUE)
emodnet_map_logo(plot, path = paste0(tempfile(), ".png"),
width = 198, height = 121, dpi = 300, units = "mm")
}
