% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.r
\docType{data}
\name{world}
\alias{world}
\title{Medium high resolution world countries.}
\format{
A sf data.frame
}
\source{
\url{https://github.com/ropensci/rnaturalearth}
}
\usage{
world
}
\description{
A dataset containing polygons of the countries of the world,
retrieved as described in ./data-raw/world.r
}
\keyword{datasets}
