% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_map_basic.R
\name{emodnet_map_basic}
\alias{emodnet_map_basic}
\title{Creates a ggplot map using EMODnet colors}
\usage{
emodnet_map_basic(
  seaColor = emodnet_colors()$darkgrey,
  landColor = emodnet_colors()$lightgrey,
  crs = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{seaColor}{character HEX code color}

\item{landColor}{character HEX code color}

\item{crs}{integer or characnter. EPGS code or projection linestring}

\item{xlim}{vector with minimum and maximum longitude}

\item{ylim}{vector with minimum and maximum latitude}
}
\value{
basic ggplot2 map
}
\description{
Creates a ggplot map using EMODnet colors
}
\examples{
emodnet_map_basic()
emodnet_map_basic(crs = 3035, xlim = c(2426378, 7093974), ylim = c(1308101, 5446513))
}
