% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_map_logo.R
\name{emodnet_map_logo}
\alias{emodnet_map_logo}
\title{Add the EMODnet logo to your ggplot}
\usage{
emodnet_map_logo(
  plot,
  path = NULL,
  gravity = "northeast",
  offset = "+0+0",
  color = TRUE,
  ...
)
}
\arguments{
\item{plot}{ggplot}

\item{path}{character}

\item{gravity}{character}

\item{offset}{character}

\item{color}{boolean}

\item{...}{params to be passed to ggplot2::ggsave()}
}
\value{
png
}
\description{
Add the EMODnet logo to your ggplot
}
\examples{
data <- eurobis::getEurobisGrid(aphiaid = 107451, gridsize = '30m')
plot <- emodnet_map_plot(data, fill = data$RecordCount,
title = "Eriocheir sinensis", subtitle = "107451", legend = "Abundance")
emodnet_map_logo(plot, path = paste0(tempfile(), ".png"),
width = 198, height = 121, dpi = 300, units = "mm")
}
