% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_map_plot.R
\name{emodnet_map_plot}
\alias{emodnet_map_plot}
\title{Plot your spatial data following the EMODnet style}
\usage{
emodnet_map_plot(
  data,
  fill = NULL,
  title = NULL,
  subtitle = NULL,
  legend = NULL,
  crs = 3035,
  xlim = c(2426378.0132, 7093974.6215),
  ylim = c(1308101.2618, 5446513.5222),
  direction = 1,
  option = "viridis",
  zoom = FALSE,
  plot_polygon_border = TRUE,
  ...
)
}
\arguments{
\item{data}{sf, RasterLayer, "SpatialPolygonsDataFrame", "SpatialPointsDataFrame", "SpatialLinesDataFrame"}

\item{fill}{vector}

\item{title}{character}

\item{subtitle}{character}

\item{legend}{character}

\item{crs}{integer}

\item{xlim}{vector}

\item{ylim}{vector}

\item{direction}{integer}

\item{option}{character}

\item{zoom}{boolean}

\item{plot_polygon_border}{boolean}

\item{...}{params to be passed to emodnet_map_basic}
}
\value{
ggplot
}
\description{
Plot your spatial data following the EMODnet style
}
\examples{
aphiaid <- 107451
specname <- "Eriocheir sinensis"
Esgrid <- sf::st_read(paste0("http://geo.vliz.be/geoserver/wfs/ows?", "service=WFS&version=1.3.0&",
"request=GetFeature&", "typeName=Dataportal\%3Aeurobis_grid_1d",
"-obisenv&", "viewParams=aphiaid\%3A", aphiaid, "&", "outputFormat=json", "&maxFeatures=10"))
emodnet_map_plot(Esgrid, fill = Esgrid$RecordCount,  title = specname,
subtitle = paste("Aphiaid =",aphiaid), legend = "Abundance",plot_polygon_border=TRUE)
}
